// types.ts
export interface Project {
  title: string;
  tech: string[];
  description: string[];
  impact: string[];
}

export interface Experience {
  role: string;
  company: string;
  location: string;
  period: string;
  points: string[];
}

export interface SkillCategory {
  category: string;
  items: string[];
}